Split `librustc/traits/error_reporting.rs`
